Another huge commit
Removed more gmpack dependency
Fixed small issues

+++ buildall.sh
@@ -5,34 +5,26 @@ echo "****************************************************************"
 echo "*                        SCRIPT STARTED                        *"
 echo "****************************************************************"
 echo ""
-echo "******************* Updating from GitHub (DISCARDING ALL LOCAL CHANGES)..."
-git fetch --all
-git reset --hard origin/main
-git pull origin main
-echo ""
 
-echo "******************* Building gmpack-forwarder..."
-cd gmpack-forwarder
+if [[ $# -gt 0 ]] ; then
+    echo "******************* Updating from GitHub (DISCARDING ALL LOCAL CHANGES)..."
+    git fetch --all
+    git reset --hard origin/main
+    git pull origin main
+    echo ""
+fi
+
+echo "******************* Building forwarder..."
+cd app-forwarder
 make
 cd ..
 echo ""
 
-#echo "******************* Building gmpack-rcm..."
-#cd gmpack-rcm
-#make
-#cd ..
-#echo ""
-
-echo "******************* Building gmpack-updater..."
+echo "******************* Building application..."
 make
 echo ""
 
-echo "******************* Creating the release ZIP..."
-mkdir switch
-mkdir ./switch/gmpack-updater
-cp gmpack-updater.nro ./switch/gmpack-updater
-zip -r -9 ./gmpack-updater.zip ./switch/
-rm -rf switch
+./release.sh
 
 echo ""
